disabling ctrl+~ by default
Double ctrl+~ previously opened the client/server REPL, but it requires
log_to_console to be set to at least ["client", "server"] (which is turned off
by default) and it barely works, because the supporting code is just a crapload
of hacks bunched together.

It's still useful for hacking, but much less so for a normal user, especially
since we can now log stuff to file. Hence I'm disabling it by default.