Made the plugin more user-friendly
"Ensime: Startup" is now always available and upon possible errors
(no project, no .ensime file, bad .ensime file) guides the user
throw a bunch of dialogs and fixes the problems (prompts for a project root,
creates a mock .ensime file with comments and pre-initialized values,
opens the offending .ensime file, so that it can be fixed quickly).

Also the .ensime file is now parsed in relaxed mode. First of all
we now allow emacs-like ";; ..." comments (that last till the end of line).
Secondly sexp.read doesn't let the input string to be padded with whitespaces,
and we do (only for parsing .ensime files, not for swank).

Finally I've removed install-ensime-server.sh, because that'd be a hassle
to keep it updated w.r.t Ensime releases.