Fixes issue #37
Instead of triggering typechecking during on_load, it does it
on_activation. Now, we don't want to do it on each activation
(which is triggered each time we switch to the tab) -- so we
keep a set of typechecked files in the env.

One thing I don't understand is why is ensime typechecking all
the previously-opened files instead of only the opened ones...